requirements.txt
httplib2==0.9.2
oauth2==1.9.0.post1
PyYAML==3.11
retrying==1.3.3
six==1.10.0
wheel==0.24.0
yelp==1.0.2
